+++ src/main/java/ar/edu/unlam/tallerweb1/domain/envio/ServicioEnvio.java
@@ -12,7 +12,7 @@ public interface ServicioEnvio {
 
     //List<Contenedor> getContenedores();
 
-    Double distanciaEnvio();
+    Integer distanciaEnvio();
 
     Double calcularCostoEnvio(Vehiculo vehiculoAsignado);
